+++ src/main/java/com/chucknorris/game/GameResponse.java
@@ -1,0 +1,4 @@
+package com.chucknorris.game;
+
+public class GameResponse {
+}
